# sql: wire up the ported index constraints code