--- a/pkg/sql/opt/idxconstraint/spans.go
+++ b/pkg/sql/opt/idxconstraint/spans.go
@@ -163,8 +163,6 @@ func ExactPrefix(spans LogicalSpans, evalCtx *tree.EvalContext) int {
 	}
 }
 
-var _ = ExactPrefix
-
 // IndexColumnInfo encompasses the information for index columns, needed for
 // index constraints.
 type IndexColumnInfo struct {