#### opt: replace placeholders with values

If we have value for placeholders, replace them with their values
during build.

Release note: None

#### opt: prefer short cast syntax

We are building CastExpr using the explicit syntax. This causes some
churn in some distsql tests when the new index constraints code is on.
Switching to the short syntax.

Release note: None

#### sql: wire up the ported index constraints code

The new index constraints code was ported to the new optimizer
structures (in `opt/idxconstraint`). This change rewires sql to use
that.

Planning benchmark differences below. Note that we will be looking at
improving the performance before 2.1 ships.

```
name                                                             old time/op    new time/op    delta
Planning/SELECT_*_FROM_abc-8                                        265µs ± 3%     266µs ± 3%     ~     (p=0.886 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-8                 200µs ± 1%     214µs ± 1%   +6.88%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_b_=_5-8                            196µs ± 2%     213µs ± 5%   +8.59%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-8                 211µs ± 4%     225µs ± 2%   +6.54%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_c_=_5-8                            195µs ± 3%     209µs ± 4%   +6.85%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-8     459µs ± 4%     435µs ± 4%     ~     (p=0.057 n=4+4)

name                                                             old alloc/op   new alloc/op   delta
Planning/SELECT_*_FROM_abc-8                                       33.8kB ± 2%    33.8kB ± 2%     ~     (p=1.000 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-8                23.4kB ± 2%    27.1kB ± 1%  +15.86%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_b_=_5-8                           25.1kB ± 1%    28.9kB ± 2%  +15.26%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-8                26.2kB ± 3%    29.7kB ± 0%  +13.19%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_c_=_5-8                           24.9kB ± 0%    28.7kB ± 1%  +15.48%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-8    70.0kB ± 2%    69.6kB ± 0%     ~     (p=0.886 n=4+4)

name                                                             old allocs/op  new allocs/op  delta
Planning/SELECT_*_FROM_abc-8                                          354 ± 0%       354 ± 0%     ~     (p=1.000 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_a_>_5_ORDER_BY_a-8                   328 ± 0%       372 ± 0%  +13.41%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_b_=_5-8                              320 ± 0%       364 ± 0%  +13.74%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_b_=_5_ORDER_BY_a-8                   360 ± 0%       404 ± 0%  +12.14%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_WHERE_c_=_5-8                              317 ± 0%       361 ± 0%  +13.96%  (p=0.029 n=4+4)
Planning/SELECT_*_FROM_abc_JOIN_abc_AS_abc2_ON_abc.a_=_abc2.a-8       649 ± 0%       648 ± 0%     ~     (p=0.429 n=4+4)
```

Release note: None