@knz 2018-02-28
First two commits LGTM modulo resolution of my worry for the 2nd one. 

I don't fully understand the 3rd commit: is that the new default path for all queries?

---

Reviewed 1 of 1 files at r1, 2 of 2 files at r2, 3 of 3 files at r3.
Review status: all files reviewed at latest revision, 2 unresolved discussions, some commit checks pending.

---

*[pkg/sql/opt/optbuilder/scalar.go, line 176 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-L6Sz6z3KiRNZaz99DOY:-L6Sz6z3KiRNZaz99DOZ:b-i830k1) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c630a15e6b081f92a6974984e05cdf1ee9e3785/pkg/sql/opt/optbuilder/scalar.go#L176)):*
> ```Go
> 			}
> 			if _, ok := d.(*tree.Placeholder); ok {
> 				panic(errorf("no placeholder value"))
> ```

this case seems impossible. `(*Placeholder).Eval()` never returns the original placeholder. Or did you only mean it as an assertion? Then maybe use `programming error` in the error text.

---

*[pkg/sql/sem/tree/expr.go, line 1262 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-L6SzUJfo3I2lwtkOqEC:-L6SzUJfo3I2lwtkOqED:b-60bkov) ([raw file](https://github.com/cockroachdb/cockroach/blob/b66dc20e7db95161f6028844ee4d5f522fadaf1e/pkg/sql/sem/tree/expr.go#L1262)):*
> ```Go
> 		return nil, err
> 	}
> 	node := &CastExpr{Expr: expr, Type: colType, SyntaxMode: CastShort}
> ```

Are you sure that's safe? W.r.t operator precedence? e.g. `CAST(a+b AS INT)` turned into `a + b::INT` or the same with some of the other operators? 

IIRC we don't have sufficient tests to exercise this and we have been bitten before.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-:-L6T-UuUrJjlyNWnPemu:b-o8szbg)*
<!-- Sent from Reviewable.io -->