@RaduBerinde 2018-02-28
Yes, the `opt/index_constraints.go` code is defunct as of this PR. I plan to delete `sql/opt/*.go` and move `sql/opt/opt/*` to `sql/opt/` in a follow-up.

---

Review status: all files reviewed at latest revision, 2 unresolved discussions, some commit checks pending.

---

*[pkg/sql/opt/optbuilder/scalar.go, line 176 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-L6Sz6z3KiRNZaz99DOY:-L6T01dKUVGqUxWet-ZN:btncoi0) ([raw file](https://github.com/cockroachdb/cockroach/blob/6c630a15e6b081f92a6974984e05cdf1ee9e3785/pkg/sql/opt/optbuilder/scalar.go#L176)):*
<details><summary><i>Previously, knz (kena) wrote…</i></summary><blockquote>

this case seems impossible. `(*Placeholder).Eval()` never returns the original placeholder. Or did you only mean it as an assertion? Then maybe use `programming error` in the error text.
</blockquote></details>

Removed (it does return it if `!HasPlaceholders()`, but I already checked that)

---

*[pkg/sql/sem/tree/expr.go, line 1262 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-L6SzUJfo3I2lwtkOqEC:-L6T1QmF18Iul5qLdVrc:bfhhfa1) ([raw file](https://github.com/cockroachdb/cockroach/blob/b66dc20e7db95161f6028844ee4d5f522fadaf1e/pkg/sql/sem/tree/expr.go#L1262)):*
<details><summary><i>Previously, knz (kena) wrote…</i></summary><blockquote>

Are you sure that's safe? W.r.t operator precedence? e.g. `CAST(a+b AS INT)` turned into `a + b::INT` or the same with some of the other operators? 

IIRC we don't have sufficient tests to exercise this and we have been bitten before.
</blockquote></details>

I added a testcase. `CAST(a + b + c AS string)` becomes `((a + b) + c)::STRING`.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-:-L6T1eqHvcrLlLPPyjGb:b-e9gkf6)*
<!-- Sent from Reviewable.io -->