@jordanlewis 2018-03-01
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 1 of 6 files at r4, 2 of 2 files at r5, 3 of 3 files at r6.
Review status: all files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---

*[pkg/sql/opt_index_selection.go, line 146 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-L6XYtqNn87qD45HhPwS:-L6XYtqNn87qD45HhPwT:b-i0k7f8) ([raw file](https://github.com/cockroachdb/cockroach/blob/8a4d2840c50dc3ee863c11d33a9ac2f6016da297/pkg/sql/opt_index_selection.go#L146)):*
> ```Go
> 			ctx, &p.semaCtx, p.EvalContext(), optimizer.Factory(), colNames, colTypes,
> 		)
> 		bld.AllowUnsupportedExpr = true
> ```

What's this about? I'm definitely unfamiliar with this code, but any time there's an uncommented boolean switch set for something it makes me nervous that future maintainers will be confused about what's going on and possibly screw up. For example, someone who wants to copy this code to support this same thing in a different area might blindly copy this flag without knowing what it does, etc etc.

Point is, can you add a comment?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-:-L6XaD3OejCA5SB9iSbD:bnfp4nl)*
<!-- Sent from Reviewable.io -->