@knz 2018-03-01
Code LGTM

Please have someone from the perf team (Jordan for example) acknowledge
the perf hit. That will need to be known when folk will start measuring
/ comparing perf across branches and over time.