@jordanlewis 2018-03-01
Yeah @knz actually makes a pretty good point... we need to communicate that we shouldn't be testing perf stuff against master anymore at all. I will make sure to share that around.