@RaduBerinde 2018-03-01
Review status: all files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---

*[pkg/sql/opt_index_selection.go, line 146 at r6](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236#-L6XYtqNn87qD45HhPwS:-L6Xcbw1-z4iaAC0wVWW:bpiaw71) ([raw file](https://github.com/cockroachdb/cockroach/blob/8a4d2840c50dc3ee863c11d33a9ac2f6016da297/pkg/sql/opt_index_selection.go#L146)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

What's this about? I'm definitely unfamiliar with this code, but any time there's an uncommented boolean switch set for something it makes me nervous that future maintainers will be confused about what's going on and possibly screw up. For example, someone who wants to copy this code to support this same thing in a different area might blindly copy this flag without knowing what it does, etc etc.

Point is, can you add a comment?
</blockquote></details>

It is documented here: https://github.com/cockroachdb/cockroach/blob/6e197ccdad6baf99fd7e6a9249fc29375a00236e/pkg/sql/opt/optbuilder/builder.go#L49

I don't think we should replicate that everywhere we use it. It will go away at some point anyway.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/23236)*
<!-- Sent from Reviewable.io -->